# Conflicts:
#	app/src/main/java/com/ch2ps215/mentorheal/di/RemoteModule.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/navgraph/NavGraph.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/navgraph/Route.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/tracker/DetailTrackerScreen.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/tracker/TrackerScreen.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/twos/DetectionPagingSource.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/twos/TwosScreen.kt
#	app/src/main/java/com/ch2ps215/mentorheal/presentation/twos/TwosViewModel.kt

+++ app/src/main/java/com/ch2ps215/mentorheal/data/remote/payload/SaveDetectionFormRequest.kt
@@ -3,6 +3,6 @@ package com.ch2ps215.mentorheal.data.remote.payload
 data class SaveDetectionFormRequest(
     val id: String,
     val label: String,
-    val scores: Float,
+    val scores: Int,
     val idUser: String,
 )
